[CLN] (test_)mail, portal: remove useless test class
We may have a small "not modular" hack in mail if it avoid having
a new class just to generate sign parameters located in portal module.

Task-4332801 [test_mail] Cleanup controller tests
Prepares Task-4332797 : [mail] Partner from emails 3.0
Prepares Task-4273479 : [mail] Email-like recipients
