+++ addons/portal/tests/__init__.py
@@ -3,6 +3,5 @@
 from . import test_login
 from . import test_message_format_portal
 from . import test_portal
-from . import test_portal_controller_common
 from . import test_tours
 from . import test_portal_wizard
