+++ addons/test_mail_full/tests/test_controller_attachment.py
@@ -2,11 +2,10 @@
 
 import odoo
 from odoo.addons.mail.tests.common_controllers import MailControllerAttachmentCommon
-from odoo.addons.portal.tests.test_portal_controller_common import TestPortalControllerCommon
 
 
 @odoo.tests.tagged("-at_install", "post_install", "mail_controller")
-class TestPortalAttachmentController(MailControllerAttachmentCommon, TestPortalControllerCommon):
+class TestPortalAttachmentController(MailControllerAttachmentCommon):
     def test_attachment_upload_portal(self):
         """Test access to upload an attachment on portal"""
         record = self.env["mail.test.portal.no.partner"].create({"name": "Test"})
