+++ addons/test_mail_full/tests/test_controller_update.py
@@ -1,10 +1,9 @@
 from odoo.addons.mail.tests.common_controllers import MailControllerUpdateCommon
-from odoo.addons.portal.tests.test_portal_controller_common import TestPortalControllerCommon
 from odoo.tests import tagged
 
 
 @tagged("-at_install", "post_install", "mail_controller")
-class TestPortalMessageUpdateController(MailControllerUpdateCommon, TestPortalControllerCommon):
+class TestPortalMessageUpdateController(MailControllerUpdateCommon):
     def test_message_update_portal(self):
         """Test Only Admin and Portal User can update a portal user message on a record with no assigned partner."""
         record = self.env["mail.test.portal.no.partner"].create({"name": "Test"})
